+++ libs/compiler/compiler.c
@@ -38,7 +38,7 @@
 #define FILE_DEBUG
 
 
-static void process_user_requests(compiler_context *context, const workspace *const ws)
+static void process_user_requests(compiler_context *context, workspace *const ws)
 {	
 #if !defined(FILE_DEBUG) && !defined(_MSC_VER)
 	/* Regular file */
@@ -88,7 +88,7 @@ static void process_user_requests(compiler_context *context, const workspace *co
 	free(macro_processed);
 }
 
-int compile_to_vm(const workspace *const ws)
+int compile_to_vm(workspace *const ws)
 {
 	if (!ws_is_correct(ws))
 	{
