+++ libs/preprocessor/macros_get.h
@@ -24,8 +24,6 @@ extern "C" {
 #endif
 
 void define_get_from_macrotext(int r, preprocess_context *context);
-void define_relis(preprocess_context *context);
-void set_relis(preprocess_context *context);
 
 #ifdef __cplusplus
 } /* extern "C" */
