+++ libs/preprocessor/parser.h
@@ -1,5 +1,5 @@
 /*
- *	Copyright 2020 Andrey Terekhov, Egor Anikin
+ *	Copyright 2018 Andrey Terekhov, Egor Anikin
  *
  *	Licensed under the Apache License, Version 2.0 (the "License");
  *	you may not use this file except in compliance with the License.
@@ -18,12 +18,11 @@
 
 #include "context_var.h"
 
-
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-void if_relis(preprocess_context *context);
+void preprocess_scan(preprocess_context *context);
 
 #ifdef __cplusplus
 } /* extern "C" */
