+++ libs/preprocessor/preprocessor.h
@@ -25,9 +25,6 @@
 extern "C" {
 #endif
 
-void preprocess_scan(preprocess_context *context);
-
-
 /**
  *	Preprocess files from workspace
  *
